Change logic for medusa_*_last_* metrics calculation.

+++ medusa_collector/medusa_struct.go
@@ -24,7 +24,7 @@ type backup struct {
 	MissingNodesList    []string `json:"missing_nodes_list"`
 	Name                string   `json:"name"`
 	Nodes               []node   `json:"nodes"`
-	NumObjects          int      `json:"num_objects"`
+	NumObjects          int64    `json:"num_objects"`
 	Size                int64    `json:"size"`
 	Started             int64    `json:"started"`
 }
@@ -41,7 +41,7 @@ type backup struct {
 type node struct {
 	Finished       int64  `json:"finished"`
 	FQDN           string `json:"fqdn"`
-	NumObjects     int    `json:"num_objects"`
+	NumObjects     int64  `json:"num_objects"`
 	ReleaseVersion string `json:"release_version"`
 	ServerType     string `json:"server_type"`
 	Size           int64  `json:"size"`
